macro: Move ALIGN_TO to macro-fundamental.h and introduce CONST_ALIGN_TO
